// src/app/interfaces/completionResponse.ts
// Generated by https://quicktype.io

export interface CompletionResponseItems {
    label:     string;
    iconclass: Iconclass;
    html:      string;
}

export enum Iconclass {
    SlIconTypeBus = "sl-icon-type-bus",
    SlIconTypeTrain = "sl-icon-type-train",
    SlIconTypeTram = "sl-icon-type-tram",
}

export type CompletionResponse = CompletionResponseItems []; 