// src/app/interfaces/routeResponse.ts
// Generated by https://quicktype.io

export interface RouteResponse {
    count:        number;
    min_duration: number;
    rawtime:      number;
    max_duration: number;
    maxtime:      number;
    connections:  Connection[];
    url:          string;
    points:       Point[];
    description:  string;
    request:      null;
    eof:          number;
}

export interface Connection {
    duration:     number;
    from:         From;
    departure:    Date;
    to:           From;
    arrival:      Date;
    is_main?:     boolean;
    disruptions?: Disruptions
    occupancy:    string;
    legs:         Leg[];
}

export interface Disruptions {
    "ch:1:sstid:100001:7c8d5f00-e742-4fdc-a53f-0fac9e2d5aca_1": Ch1Sstid1000017c8d5f00E7424fdcA53f0fac9e2d5aca1
}

export interface Ch1Sstid1000017c8d5f00E7424fdcA53f0fac9e2d5aca1 {
    id: string
    perspective: string[]
    periods: Periods
    texts: Texts
    Progress: string
    AlertCause: string
    Priority: string
    type: string
  }

export interface Periods {
    validity:    Array<number[]>;
    publication: Array<number[]>;
}

export interface Texts {
    S: S
    M: M
    L: L
}

export interface S {
    summary: string
    reason: string
    duration: string
    recommendation: string
    consequence: string
  }
  
  export interface M {
    summary: string
    reason: string
    duration: string
    recommendation: string
    consequence: string
  }
  
  export interface L {
    summary: string
    reason: string
    duration: string
    recommendation: string
    consequence: string
  }
export enum From {
    Genève = "Genève",
    Lausanne = "Lausanne",
}

export interface Leg {
    departure?: Date
    x: any
    y: any
    isaddress?: boolean
    name: string
    sbb_name: string
    stopid: string
    type?: string
    terminal?: string
    runningtime?: number
    exit?: Exit
    type_name?: string
    lon: number
    lat: number
    arrival?: string
    tripid?: string
    line?: string
    waittime?: number
    fgcolor?: string
    bgcolor?: string
    "*G"?: string
    "*L"?: string
    "*Z"?: string
    operator?: string
    stops?: Stop[]
    contop_stop: any
    occupancy?: string
    track?: string
    attributes?: Attributes
    disruptions: any
    normal_time?: number
}

export interface Attributes {
    "0_1.4_WR"?: string;
    "0_4.8_RZ"?: string;
    "0_8.1_R"?:  string;
    "1_4.6_FA"?: string;
    "1_4.7_BZ":  string;
    "1_4.7_FS":  string;
    "1_4.6_FZ"?: string;
}

export interface Exit {
    arrival:    Date;
    stopid:     string;
    x:          number;
    y:          number;
    name:       string;
    sbb_name?:  From;
    track?:     string;
    lon:        number;
    lat:        number;
    departure?: Date;
    occupancy?: string;
}

export interface Stop {
    arrival: string
    departure: string
    occupancy: string
    name: string
    stopid: string
    x: number
    y: number
    lon: number
    lat: number
  }

export interface Point {
    id:   string;
    text: From;
    url:  string;
    x:    string;
    y:    string;
    lon:  number;
    lat:  number;
}